stepSize -> Integer

+++ services/recom-backend/src/main/java/com/recom/dto/map/scanner/topography/MapTopographyEntityDto.java
@@ -23,7 +23,7 @@ public class MapTopographyEntityDto implements MapRelatedDto {
     // META DATA
     @Schema
     @JsonProperty()
-    private Float stepSize;
+    private Integer stepSize;
 
     @Schema
     @JsonProperty()
